refactor(enrollment) : 포트/서비스 구조 정리 및 리팩토링 추가
