+++ src/main/java/com/lxp/aplus/enrollment/application/port/in/EnrollmentCommandPort.java
@@ -1,0 +1,10 @@
+package com.lxp.aplus.enrollment.application.port.in;
+
+import com.lxp.aplus.enrollment.application.command.EnrollmentCommand;
+import com.lxp.aplus.enrollment.domain.Enrollment;
+
+public interface EnrollmentCommandPort {
+    Long enroll(EnrollmentCommand command);
+
+    Enrollment cancel(Long enrollmentId, Long studentId);
+}
