+++ src/main/java/com/lxp/aplus/enrollment/application/service/EnrollmentQueryService.java
@@ -1,9 +1,10 @@
-package com.lxp.aplus.enrollment.application.usecase;
+package com.lxp.aplus.enrollment.application.service;
 
 import com.lxp.aplus.common.error.BusinessException;
 import com.lxp.aplus.common.error.code.CourseErrorCode;
 import com.lxp.aplus.common.error.code.EnrollmentErrorCode;
-import com.lxp.aplus.enrollment.application.port.out.CourseFinder;
+import com.lxp.aplus.enrollment.application.port.in.EnrollmentQueryPort;
+import com.lxp.aplus.enrollment.application.port.out.CourseReader;
 import com.lxp.aplus.enrollment.application.port.out.CourseSummary;
 import com.lxp.aplus.enrollment.application.port.out.ProgressQueryPort;
 import com.lxp.aplus.enrollment.application.port.out.dto.EnrollmentProgressDto;
@@ -25,22 +26,22 @@ import java.util.stream.Collectors;
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
-public class EnrollmentQueryUseCase {
+public class EnrollmentQueryService implements EnrollmentQueryPort {
 
     private final EnrollmentRepository enrollmentRepository;
-    private final CourseFinder courseFinder;
+    private final CourseReader courseReader;
     private final ProgressQueryPort progressQueryPort;
 
     public Page<EnrollmentListItemResult> getEnrollmentList(Long studentId, EnrollmentStatus status, Pageable pageable) {
         Page<Enrollment> enrollmentsPage = enrollmentRepository.findByStudentIdAndStatus(studentId, status, pageable);
 
-        // TODO: [성능 개선] N+1 쿼리 발생 지점. CourseFinder에 findCourseByIds(List<Long> courseIds)와 같은 배치 조회 기능 추가 필요.
+        // TODO: [성능 개선] N+1 쿼리 발생 지점. CourseReader에 findCoursesByIds(List<Long> courseIds)와 같은 배치 조회 기능 추가 필요.
         List<EnrollmentListItemResult> content = enrollmentsPage.getContent().stream()
                 .map(enrollment -> {
-                    CourseSummary courseSummary = courseFinder.findCourseById(enrollment.getCourseId())
+                    CourseSummary courseSummary = courseReader.findCourseById(enrollment.getCourseId())
                             .orElseThrow(() -> new BusinessException(CourseErrorCode.COURSE_NOT_FOUND));
                     
-                    List<String> categoryNames = courseFinder.findCategoryNamesByCourseId(enrollment.getCourseId());
+                    List<String> categoryNames = courseReader.findCategoryNamesByCourseId(enrollment.getCourseId());
 
                     EnrollmentProgressDto progress = progressQueryPort.getProgress(enrollment.getId(), enrollment.getCourseId());
